+++ runtime/test/unit/runner.clj
@@ -30,6 +30,7 @@
             [ujima.linux.audio-test]
             [ujima.linux.i3-test]
             [ujima.linux.net-test]
+            [ujima.linux.net.mdns-test]
             [ujima.linux.net.wifi-test]
             [ujima.linux.systemd-test]
             [ujima.linux.disk.block-test]
@@ -77,6 +78,7 @@
     ujima.linux.audio-test
     ujima.linux.i3-test
     ujima.linux.net-test
+    ujima.linux.net.mdns-test
     ujima.linux.net.wifi-test
     ujima.linux.systemd-test
     ujima.linux.disk.block-test